+++ Content/UI/DriveBayUISystem.cs
@@ -185,5 +185,6 @@ namespace TerraStorage.Content.UI
                 }
             }
         }
+
     }
 }